% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_poly.R
\name{load_preprocess_data}
\alias{load_preprocess_data}
\title{Load and preprocess input data}
\usage{
load_preprocess_data(path_biomes)
}
\arguments{
\item{path_biomes}{Path to the input shapefile containing the biomes}
}
\value{
A list of sf objects, each representing a polygon
}
\description{
Load and preprocess input data
}
