% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{Post-process a raster file}
\usage{
post_process(
  input_file,
  output_file,
  reference_crs,
  reference_pixel_size,
  temp_dir,
  resampling_method = "near"
)
}
\arguments{
\item{input_file}{Character. The path to the input raster file.}

\item{output_file}{Character. The path where the processed raster file will be saved.}

\item{reference_crs}{Character. The target coordinate reference system (CRS) in Well-Known Text (WKT) format.}

\item{reference_pixel_size}{Numeric vector of length 2. The target pixel size (resolution) in the x and y directions.}

\item{temp_dir}{Character. The path to the directory where temporary files will be stored.}

\item{resampling_method}{Character. The resampling method to use for reprojecting and aligning the raster. Options are "near", "bilinear", "cubic", "cubicspline", "lanczos", "average", "mode", "max", "min", "med", "q1", "q3". Default is "near".}
}
\value{
NULL. The function is called for its side effect of saving the processed raster to the specified output file.
}
\description{
This function post-processes a raster file by changing the data type, compressing, reprojecting, aligning, and trimming it.
}
\details{
This function performs the following steps:
\itemize{
\item Changes the data type of the raster and compresses it.
\item Reprojects and aligns the raster to the specified CRS and pixel size.
\item Trims the raster to remove outer rows and columns that are all NA values.
}
Temporary files are created during the process and deleted upon completion. If the raster contains only NA values after trimming, it is skipped and not processed further.
}
\examples{
\dontrun{
post_process(
  input_file = "path/to/input.tif",
  output_file = "path/to/output.tif",
  reference_crs = "EPSG:4326",
  reference_pixel_size = c(0.00025, 0.00025),
  temp_dir = "path/to/temp",
  resampling_method = "bilinear"
)
}
}
\author{
Jeronimo Rodriguez-Escobar \href{mailto:jeronimo.rescobar@gmail.com}{jeronimo.rescobar@gmail.com}
}
