% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc _all.R
\name{process_sublists}
\alias{process_sublists}
\title{Process each sublist of sf objects to download and save rasters}
\usage{
process_sublists(biomat, output_dir, download_path, n_cores = 2)
}
\arguments{
\item{biomat}{A list of sublists of sf objects}

\item{output_dir}{The directory where the output rasters will be saved}

\item{download_path}{The path where downloaded rasters will be stored}

\item{n_cores}{Number of cores for parallel processing}
}
\value{
None
}
\description{
Process each sublist of sf objects to download and save rasters
}
