% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_list.R
\name{split_list}
\alias{split_list}
\title{Split a list into sublists}
\usage{
split_list(input_list, n)
}
\arguments{
\item{input_list}{The input list to be split}

\item{n}{The number of sublists}
}
\value{
A list of sublists
}
\description{
Split a list into sublists
}
